# ecis-core/src/main/java/cz/ecis/core/model/LovDto.java
package cz.ecis.core.model;

import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
public class LovDto {
    
    private Object id;
    private String label;
}